Compatible with v5.8
In 5.8, the pointer to the make_request_fn function has not yet been cut, as for example in 5.9, but it can be NULL, since it is not defined. In this case, in order to complete the original request, we can use the blk_mq_make_request function.
This solution is temporary and only for 5.8, since with 5.9 neither make_request_fn nor blk_mq_make_request is already there.